pyproject.toml
[project]
name = "text-analysis-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "google-api-python-client>=2.193.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[google,mcp]>=1.70.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
    "wikipedia-api>=0.11.0",
    "wordcloud>=1.9.6",
]
